Move VSCode extension mapping directly into Plugin creation
